feat: code structured in Routes, Controller and Middleware directory and added type safety
