+++ backend/src/middleware/errorHandlingMiddleware.ts
@@ -1,0 +1,10 @@
+import {Request,Response,NextFunction} from 'express'
+
+export const errorMiddleware = (err: any, req: Request, res: Response, next: NextFunction) => {
+    console.error(err.stack);
+    res.status(500).json({
+      error: 'Something went wrong!',
+      message: process.env.NODE_ENV === 'development' ? err.message : 'Internal Server Error'
+    });
+    next();
+}
