+++ backend/src/routes/projects.routes.ts
@@ -1,0 +1,8 @@
+import express from 'express'
+import { getProjects } from '../controllers/projects.controller';
+
+const router = express.Router();
+
+router.get("/api/projects",getProjects)
+
+export default router
