+++ backend/src/routes/stats.routes.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+import { getStats } from '../controllers/stats.controller';
+const router = express.Router();
+
+router.get("/api/stats",getStats)
+
+export default router
