+++ backend/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/index.ts","./src/controllers/contributors.controller.ts","./src/controllers/projects.controller.ts","./src/controllers/stats.controller.ts","./src/middleware/404handlingmiddleware.ts","./src/middleware/errorhandlingmiddleware.ts","./src/routes/contributors.routes.ts","./src/routes/health.routes.ts","./src/routes/projects.routes.ts","./src/routes/stats.routes.ts"],"version":"5.7.3"}
